try to build vtk wheel from source

+++ ci_wheels/build_wheels.sh
@@ -8,9 +8,18 @@ pip install wheel
 
 for pkg in ${WHEEL_PKGS} ; do
   cd /src
-  pip download --no-deps $pkg -d /src/
-  unp /src/${pkg}*.tar.gz && rm /src/${pkg}*.tar.gz
-  cd ${pkg}*
+  if [[ "${pkg}" == "vtk" ]] ; then
+    wget -qO /src/${pkg}.tar.gz https://www.vtk.org/files/release/9.0/VTK-9.0.3.tar.gz
+    unp /src/${pkg}.tar.gz
+    mkdir /tmp/build
+    cd /tmp/build
+    cmake /src/VTK-* -DVTK_WHEEL_BUILD=ON -DVTK_PYTHON_VERSION=3
+    make
+  else
+    pip download --no-binary :all: --no-deps $pkg -d /src/
+    unp /src/${pkg}*.tar.gz && rm /src/${pkg}*.tar.gz
+    cd ${pkg}*
+  fi
   pip wheel --use-feature=in-tree-build . -w ${WHEELHOUSE}/tmp
   mv ${WHEELHOUSE}/tmp/${pkg}* ${WHEELHOUSE}/
   rm -rf ${WHEELHOUSE}/tmp
